% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.efs-package.R
\docType{package}
\name{aws.efs-package}
\alias{aws.efs-package}
\alias{aws.efs}
\title{aws.efs}
\description{
AWS EFS Client Package
}
\details{
A simple client package for the Amazon Web Services (AWS) Elastic File System (EFS) REST API.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
