% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_efs.R
\name{create_efs}
\alias{create_efs}
\title{Create EFS File System}
\usage{
create_efs(mode = c("generalPurpose", "maxIO"), token = NULL, ...)
}
\arguments{
\item{mode}{A character string specifying the \dQuote{Performance Mode} for the file system. \dQuote{generalPurpose} (the default) is appropriate in most situations.}

\item{token}{An optional character string specifying a token to ensure idempotency. Must be between 1 and 64 characters in length.}

\item{\dots}{Additional arguments passed to \code{\link{efsHTTP}}.}
}
\value{
A list.
}
\description{
Create an EFS File System
}
\references{
\href{http://docs.aws.amazon.com/efs/latest/ug/API_CreateFileSystem.html}{API Documentation}
}
\seealso{
\code{\link{delete_efs}}, \code{\link{create_mount}}
}

