% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mount.R
\name{create_mount}
\alias{create_mount}
\title{Create EFS Mount Target}
\usage{
create_mount(id, ip, sgroup, subnet, ...)
}
\arguments{
\item{id}{A character string specifying an EFS File System ID, perhaps as returned by \code{\link{create_efs}} or \code{\link{efs_list}}.}

\item{ip}{An object of class \dQuote{ec2_ip} or an allocationId (for a VPC IP).}

\item{sgroup}{A character vector specifying one or more Security Group IDs, or (a list of) objects of class \dQuote{ec2_security_group}. No more than 5 security groups can be specified. These values can be modified later using \code{\link{replace_mount_sgroups}}.}

\item{subnet}{A character string containing the name of a subnet, or an object of class \dQuote{ec2_subnet}.}

\item{\dots}{Additional arguments passed to \code{\link{efsHTTP}}.}
}
\value{
A list.
}
\description{
Create an EFS Mount Target
}
\details{
This function creates a mount target for an EFS File System so that it can mounted to one or more EC2 instances. All instances within an AWS availability zone share a mount target. Mount targets are subnet-specific, so one mount target should be created for each subnet.
}
\references{
\href{http://docs.aws.amazon.com/efs/latest/ug/API_CreateMountTarget.html}{API Documentation}
}
\seealso{
\code{\link{create_efs}}, \code{\link{delete_mount}}, \code{\link{get_mount_sgroups}}
}
