% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_efs.R
\name{delete_file_system}
\alias{delete_file_system}
\title{Delete EFS File System}
\usage{
delete_file_system(id, ...)
}
\arguments{
\item{id}{A character string specifying an EFS File System ID, perhaps as returned by \code{\link{create_efs}} or \code{\link{efs_list}}.}

\item{\dots}{Additional arguments passed to \code{\link{efsHTTP}}.}
}
\value{
If successful, a logical TRUE.
}
\description{
Delete an EFS File System
}
\details{
You cannot delete a File System that is in use, meaning has any mount targets. The mount targets must be deleted first for this operation to succeed, perhaps using \code{\link{delete_mount}}.
}
\references{
\href{http://docs.aws.amazon.com/efs/latest/ug/API_DeleteFileSystem.html}{API Documentation}
}
\seealso{
\code{\link{create_efs}}, \code{\link{delete_mount}}
}
