% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_mount.R
\name{delete_mount}
\alias{delete_mount}
\title{Delete EFS Mount Target}
\usage{
delete_mount(mount, ...)
}
\arguments{
\item{mount}{A character string specifying an EFS Mount Target ID, perhaps as returned by \code{\link{create_mount}} or \code{\link{mount_list}}.}

\item{\dots}{Additional arguments passed to \code{\link{efsHTTP}}.}
}
\value{
A list.
}
\description{
Delete an EFS Mount Target
}
\references{
\href{http://docs.aws.amazon.com/efs/latest/ug/API_DeleteMountTarget.html}{API Documentation}
}
\seealso{
\code{\link{create_efs}}, \code{\link{delete_mount}}, \code{\link{mount_list}}, \code{\link{get_mount_sgroups}}
}

