% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_efs.R
\name{efs_list}
\alias{efs_list}
\title{List EFS File System(s)}
\usage{
efs_list(id = NULL, n = NULL, marker = NULL, ...)
}
\arguments{
\item{id}{Optionally, a character string specifying an EFS File System ID, perhaps as returned by \code{\link{create_efs}} to use to restrict results.}

\item{n}{Optionally, an integer specifying the number of results to return.}

\item{marker}{Optionally, a character string specifying a pagination marker returned by a previous result.}

\item{\dots}{Additional arguments passed to \code{\link{efsHTTP}}.}
}
\value{
A list.
}
\description{
List EFS File System(s)
}
\references{
\href{http://docs.aws.amazon.com/efs/latest/ug/API_DescribeFileSystems.html}{API Documentation}
}
\seealso{
\code{\link{create_efs}}, \code{\link{delete_efs}}
}

