% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{efsHTTP}
\alias{efsHTTP}
\title{API Requests}
\usage{
efsHTTP(verb = "GET", action = "/", query = list(), body = list(),
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"), version = "2015-02-01", ...)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \dQuote{GET}.}

\item{action}{A character string specifying the path to the API endpoint.}

\item{query}{A named list of query string parameters.}

\item{body}{A list of body (JSON) arguments.}

\item{region}{A character string containing the AWS region. If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. If missing, defaults to value stored in environment variable \dQuote{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an AWS Secret Access Key.  If missing, defaults to value stored in environment variable \dQuote{AWS_SECRET_ACCESS_KEY}.}

\item{version}{A character string specifying an API version. Default is \dQuote{2015-04-13}.}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
A list
}
\description{
EFS HTTP Requests
}

