% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount_sgroups.R
\name{get_mount_sgroups}
\alias{get_mount_sgroups}
\alias{replace_mount_sgroups}
\title{List or replace Mount Target security groups}
\usage{
get_mount_sgroups(mount, ...)

replace_mount_sgroups(mount, sgroup, ...)
}
\arguments{
\item{mount}{A character string specifying an EFS Mount Target ID, perhaps as returned by \code{\link{create_mount}} or \code{\link{mount_list}}.}

\item{\dots}{Additional arguments passed to \code{\link{efsHTTP}}.}

\item{sgroup}{A character vector specifying one or more Security Group IDs, or (a list of) objects of class \dQuote{ec2_security_group}. No more than 5 security groups can be specified.}
}
\value{
A list.
}
\description{
\code{get_mount_sgroups} lists security groups for an EFS Mount Target. \code{replace_mount_sgroups} replaces them.
}
\references{
\href{http://docs.aws.amazon.com/efs/latest/ug/API_DescribeMountTargetSecurityGroups.html}{API Documentation: DescribeMountTargetSecurityGroups}

\href{http://docs.aws.amazon.com/efs/latest/ug/API_ModifyMountTargetSecurityGroups.html}{API Documentation: ModifyMountTargetSecurityGroups}
}
\seealso{
\code{\link{create_mount}}, \code{\link{delete_mount}}, \code{\link{mount_list}}
}
