% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{add_tags}
\alias{add_tags}
\alias{delete_tags}
\alias{get_tags}
\title{Get/Add/Delete EFS Tags}
\usage{
add_tags(id, tags, ...)

delete_tags(id, tags, ...)

get_tags(id, ...)
}
\arguments{
\item{id}{A character string specifying an EFS File System ID, perhaps as returned by \code{\link{create_efs}} or \code{\link{efs_list}}.}

\item{tags}{For \code{add_tags}, a named character vector specifying tag keys (as names) and values to add to the EFS. For \code{delete_tags}, a character vector of tag keys to delete.}

\item{\dots}{Additional arguments passed to \code{\link{efsHTTP}}.}
}
\value{
A list.
}
\description{
Get, add, or delete tags on an EFS file system
}
\references{
\href{http://docs.aws.amazon.com/efs/latest/ug/API_CreateTags.html}{API Documentation: CreateTags}
  \href{http://docs.aws.amazon.com/efs/latest/ug/API_DescribeTags.html}{API Documentation: DescribeTags}
  \href{http://docs.aws.amazon.com/efs/latest/ug/API_DeleteTags.html}{API Documentation: DeleteTags}
}
\seealso{
\code{\link{create_efs}}
}

