% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{efsHTTP}
\alias{efsHTTP}
\title{API Requests}
\usage{
efsHTTP(verb = "GET", action = "/", query = NULL, headers = list(),
  body = NULL, verbose = getOption("verbose", FALSE),
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"), key = NULL,
  secret = NULL, session_token = NULL, ...)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \dQuote{GET}.}

\item{action}{A character string specifying the path to the API endpoint.}

\item{query}{A named list of query string parameters.}

\item{headers}{A list of headers to pass to the HTTP request.}

\item{body}{A list of body (JSON) arguments.}

\item{verbose}{A logical indicating whether to be verbose. Default is given by \code{options("verbose")}.}

\item{region}{A character string containing the AWS region. If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. See \code{\link[aws.signature]{locate_credentials}}.}

\item{secret}{A character string containing an AWS Secret Access Key. See \code{\link[aws.signature]{locate_credentials}}.}

\item{session_token}{A character string containing an AWS Session Token. See \code{\link[aws.signature]{locate_credentials}}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
A list
}
\description{
EFS HTTP Requests
}
